Algebra.Algorithms.Faugere4

F_4 algorithms with various backends

faugere4LM :: (Eq r, Field r, KnownNat n, IsMonomialOrder n ord) => Strategy r ord n -> Ideal (OrderedPolynomial r ord n) -> Ideal (OrderedPolynomial r ord n)

F_4 with linked list based matrix with modified matrix construction

faugere4 :: (Normed r, Fractional r, IsMonomialOrder n ord, CoeffRing r, KnownNat n, Elt r, Target (DefVec r) r, Source (DefVec r) r) => Strategy r ord n -> Ideal (OrderedPolynomial r ord n) -> Ideal (OrderedPolynomial r ord n)

F_4 using repa

faugere4G :: forall r ord mat n. (Normed r, Field r, IsMonomialOrder n ord, CoeffRing r, KnownNat n, Matrix mat, Elem mat r) => Proxy mat -> Strategy r ord n -> Ideal (OrderedPolynomial r ord n) -> Ideal (OrderedPolynomial r ord n)

F_4 with general matrix.

faugere4Modular :: (KnownNat n, IsMonomialOrder n ord) => Strategy (Fraction Integer) ord n -> Ideal (OrderedPolynomial (Fraction Integer) ord n) -> Ideal (OrderedPolynomial (Fraction Integer) ord n)

Selection strategies

optimalStrategy :: Strategy r ord n

sugarStrategy :: (KnownNat n, IsMonomialOrder n ord, CoeffRing r) => Strategy r ord n

F_4 main algorithm

faugere4Gen

Arguments

:: (IsMonomialOrder n ord, CoeffRing r, KnownNat n)

=> ([OrderedPolynomial r ord n] -> (mat r, dic))
Convert list of polynomials to matrix and intermediate data.

-> (mat r -> mat r)
Gaussian elimination

-> (dic -> mat r -> [OrderedPolynomial r ord n])
Retrieve polynomials data

-> Strategy r ord n

-> Ideal (OrderedPolynomial r ord n)

-> Ideal (OrderedPolynomial r ord n)

Generate F_4 algorithm with new backend

Examples

cyclic :: KnownNat n => SNat n -> Ideal (Polynomial (Fraction Integer) n)