Algebra.Algorithms.ZeroDim

Description

Algorithms for zero-dimensional ideals.

Documentation

univPoly :: forall r ord n. (Ord r, Normed r, Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord) => Ordinal n -> Ideal (OrderedPolynomial r ord n) -> OrderedPolynomial r ord n

Calculate the monic generator of k[X_0, ..., X_n] intersect k[X_i].

radical :: forall r ord n. (Normed r, Ord r, CoeffRing r, KnownNat n, Field r, IsMonomialOrder n ord) => Ideal (OrderedPolynomial r ord n) -> Ideal (OrderedPolynomial r ord n)

Calculate the radical of the given zero-dimensional ideal.

isRadical :: forall r ord n. (Normed r, Ord r, CoeffRing r, KnownNat n, (0 :< n) ~ True, Field r, IsMonomialOrder n ord) => Ideal (OrderedPolynomial r ord n) -> Bool

Test if the given zero-dimensional ideal is radical or not.

solveWith :: forall r n ord. (DecidableZero r, Normed r, Ord r, Field r, CoeffRing r, (0 :< n) ~ True, IsMonomialOrder n ord, IsMonomialOrder (n + 1) ord, KnownNat n, Convertible r Double) => OrderedPolynomial r ord n -> Ideal (OrderedPolynomial r ord n) -> Maybe [Vector (ComplexDouble) n]

newtypeWrappedField a

Constructors

WrapField

Fields

unwrapField :: a

reduction :: (CoeffRing r, KnownNat n, IsMonomialOrder n ord, Field r) => Ordinal n -> OrderedPolynomial r ord n -> OrderedPolynomial r ord n

solveViaCompanion :: forall r ord n. (Normed r, Ord r, Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord, Convertible r Double) => Double -> Ideal (OrderedPolynomial r ord n) -> [Vector (ComplexDouble) n]

solveM :: forall m r ord n. (Normed r, Ord r, MonadRandom m, Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord, Convertible r Double, (0 :< n) ~ True, IsMonomialOrder (n + 1) ord) => Ideal (OrderedPolynomial r ord n) -> m [Sized n (ComplexDouble)]

solve' :: forall r n ord. (Field r, CoeffRing r, KnownNat n, (0 :< n) ~ True, IsMonomialOrder n ord, Convertible r Double) => Double -> Ideal (OrderedPolynomial r ord n) -> [Vector (ComplexDouble) n]

matrixRep :: (DecidableZero t, Eq t, Field t, KnownNat n, IsMonomialOrder n order, Reifies ideal (QIdeal t order n)) => Quotient t order n ideal -> [[t]]

subspMatrix :: (Normed r, Ord r, Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord) => Ordinal n -> Ideal (OrderedPolynomial r ord n) -> Matrix r

vectorRep :: forall r order ideal n. (CoeffRing r, KnownNat n, IsMonomialOrder n order, Reifies ideal (QIdeal r order n)) => Quotient r order n ideal -> Vector r

solveLinear :: (Ord r, Fractional r) => Matrix r -> Vector r -> Maybe (Vector r)

Solves linear system. If the given matrix is degenerate, this returns Nothing.

fglm :: (Normed r, Ord r, KnownNat n, Field r, IsMonomialOrder n ord, (0 :< n) ~ True) => Ideal (OrderedPolynomial r ord n) -> ([OrderedPolynomial r Lex n], [OrderedPolynomial r Lex n])

Calculate the Groebner basis w.r.t. lex ordering of the zero-dimensional ideal using FGLM algorithm. If the given ideal is not zero-dimensional this function may diverge.

fglmMap

Arguments

:: (Normed k, Ord k, Field k, (0 :< n) ~ True, IsMonomialOrder n ord, CoeffRing k, KnownNat n)

=> (OrderedPolynomial k ord n -> Vector k)
Linear map from polynomial ring.

-> ([OrderedPolynomial k Lex n], [OrderedPolynomial k Lex n])
The tuple of:

lex-Groebner basis of the kernel of the given linear map.

The vector basis of the image of the linear map.

Compute the kernel and image of the given linear map using generalized FGLM algorithm.