Algebra.Field.Galois

Documentation

data GF' p n f

Galois field of order p^n. f stands for the irreducible polynomial over F_p of degree n.

class (KnownNat n, KnownNat p, Reifies f (Unipol (F p))) => IsGF' p n f

Type-constraint synonym to work with Galois field.

modPoly :: forall p n f. (KnownNat n, Reifies p Integer) => Unipol (F p) -> GF' p n f

modVec :: Sized n (F p) -> GF' p n f

withIrreducible :: forall p a. KnownNat p => Unipol (F p) -> (forall f n. Reifies f (Unipol (F p)) => Proxy (GF' p n f) -> a) -> a

linearRepGF :: GF' p n f -> Vector (F p)

linearRepGF' :: GF' p n f -> Vector Integer

reifyGF' :: MonadRandom m => Natural -> Natural -> (forall p f n. (Reifies p Integer, Reifies f (Unipol (F p))) => Proxy (GF' p n f) -> a) -> m a

generateIrreducible :: (MonadRandom m, FiniteField k, Eq k) => proxy k -> Natural -> m (Unipol k)

generateIrreducible p n generates irreducible polynomial over F_p of degree n.

withGF' :: MonadRandom m => Natural -> Natural -> (forall p f n. (Reifies p Integer, Reifies f (Unipol (F p))) => GF' p n f) -> m (Vector Integer)

type GF p n = GF' p n (Conway p n)

Galois Field of order p^n. This uses conway polynomials as canonical minimal polynomial and it should be known at compile-time (i.e. Reifies (Conway p n) (Unipol (F n)) instances should be defined to use field operations).

class ConwayPolynomial p n where

Type-class to provide the dictionary for Conway polynomials

Minimal complete definition

conwayPolynomial

Methods

conwayPolynomial :: proxy p -> proxy n -> Unipol (F p)

data Conway p n

Empty tag to reify Conway polynomial to type-level

primitive :: forall p n f. IsGF' p n f => GF' p (n + 1) f

conway :: forall p n. ConwayPolynomial p n => SNat p -> SNat n -> Unipol (F p)

Conway polynomial (if definition is known).

conwayFile :: FilePath -> DecsQ

Parse conway polynomial file and define instances for them. File-format must be the same as Lueback's file.

addConwayPolynomials :: [(Integer, Integer, [Integer])] -> DecsQ

Macro to add Conway polynomials dictionary.