Algebra.Matrix

Documentation

classMatrix mat where

Minimal complete definition

cmap, empty, fromLists, ncols, nrows, identity, diag, getDiag, trace, diagProd, zero, colVector, rowVector, getCol, getRow, switchRows, scaleRow, combineRows, trans, buildMatrix, (!), (<||>), (<-->)

Associated Types

typeElem mat a :: Constraint

Methods

cmap :: (Elem mat a, Elem mat b) => (a -> b) -> mat a -> mat b

empty :: Elem mat b => mat b

fromLists :: Elem mat a => [[a]] -> mat a

fromCols :: Elem mat a => [Vector a] -> mat a

fromRows :: Elem mat a => [Vector a] -> mat a

toCols :: Elem mat a => mat a -> [Vector a]

toRows :: Elem mat a => mat a -> [Vector a]

ncols :: mat a -> Int

nrows :: mat a -> Int

identity :: Elem mat a => Int -> mat a

diag :: Elem mat a => Vector a -> mat a

getDiag :: Elem mat a => mat a -> Vector a

trace :: Elem mat a => mat a -> a

diagProd :: Elem mat a => mat a -> a

zero :: Elem mat a => Int -> Int -> mat a

colVector :: Elem mat a => Vector a -> mat a

rowVector :: Elem mat a => Vector a -> mat a

getCol :: Elem mat a => Int -> mat a -> Vector a

getRow :: Elem mat a => Int -> mat a -> Vector a

switchRows :: Elem mat a => Int -> Int -> mat a -> mat a

scaleRow :: Elem mat a => a -> Int -> mat a -> mat a

combineRows :: Elem mat a => Int -> a -> Int -> mat a -> mat a

trans :: Elem mat a => mat a -> mat a

buildMatrix :: Elem mat a => Int -> Int -> ((Int, Int) -> a) -> mat a

index :: Elem mat a => Int -> Int -> mat a -> Maybe a

(!) :: Elem mat a => mat a -> (Int, Int) -> a

(<||>) :: Elem mat a => mat a -> mat a -> mat a

(<-->) :: Elem mat a => mat a -> mat a -> mat a

nonZeroRows :: (DecidableZero a, Elem mat a) => mat a -> [Int]

nonZeroCols :: (DecidableZero a, Elem mat a) => mat a -> [Int]

delta :: (Monoidal r, Unital r) => Int -> Int -> r

companion :: (KnownNat n, CoeffRing r, Matrix mat, Elem mat r, IsMonomialOrder n ord) => Ordinal n -> OrderedPolynomial r ord n -> mat r

gaussReduction :: (Matrix mat, Elem mat a, Normed a, Eq a, Field a) => mat a -> (mat a, mat a)

gaussReduction a = (a', p) where a' is row echelon form and p is pivoting matrix.

maxNorm :: (Elem mat a, Normed a, Matrix mat) => mat a -> Norm a

rankWith :: (Elem mat r, CoeffRing r, Matrix mat) => (mat r -> mat r) -> mat r -> Int

det :: (Elem mat a, Eq a, Field a, Normed a, Matrix mat) => mat a -> a

inverse :: (Elem mat a, Eq a, Field a, Normed a, Matrix mat) => mat a -> mat a

inverseWith :: (mat a -> (mat a, mat a)) -> mat a -> mat a