Algebra.Prelude.Core

Documentation

(%) :: (IsPolynomial poly, Division (Coefficient poly)) => Coefficient poly -> Coefficient poly -> poly infixl 7

newtype Scalar r

Scalar r provides almost the same type-instances as r, but it can also behave as a Module over r itself.

Constructors

Scalar

Fields

runScalar :: r

(.*.) :: Module (Scalar r) m => r -> m -> m infixr 8

od :: QuasiQuoter

type Ordinal (n :: Nat) = Ordinal Nat n

enumOrdinal :: Sing Nat n -> [Ordinal n]

logBase2 :: Int -> Int

ceilingLogBase2 :: Int -> Int

module AlgebraicPrelude

module Algebra.Ring.Polynomial.Monomial

module Algebra.Ring.Polynomial.Class

type Polynomial r = OrderedPolynomial r Grevlex

transformMonomial :: (IsMonomialOrder m o, CoeffRing k, KnownNat m) => (Monomial n -> Monomial m) -> OrderedPolynomial k o n -> OrderedPolynomial k o m

castPolynomial :: (CoeffRing r, KnownNat n, KnownNat m, IsMonomialOrder n o, IsMonomialOrder m o') => OrderedPolynomial r o n -> OrderedPolynomial r o' m

changeOrder :: (CoeffRing k, Eq (Monomial n), IsMonomialOrder n o, IsMonomialOrder n o', KnownNat n) => o' -> OrderedPolynomial k o n -> OrderedPolynomial k o' n

changeOrderProxy :: (CoeffRing k, Eq (Monomial n), IsMonomialOrder n o, IsMonomialOrder n o', KnownNat n) => Proxy o' -> OrderedPolynomial k o n -> OrderedPolynomial k o' n

scastPolynomial :: (IsMonomialOrder n o, IsMonomialOrder m o', KnownNat m, CoeffRing r, KnownNat n) => SNat m -> OrderedPolynomial r o n -> OrderedPolynomial r o' m

newtype OrderedPolynomial r order n

n-ary polynomial ring over some noetherian ring R.

Constructors

Polynomial

Fields

_terms :: Map (OrderedMonomial order n) r

allVars :: forall k ord n. (IsMonomialOrder n ord, CoeffRing k, KnownNat n) => Sized n (OrderedPolynomial k ord n)

substVar :: (CoeffRing r, KnownNat n, IsMonomialOrder n ord, (1 :<= n) ~ True) => Ordinal n -> OrderedPolynomial r ord n -> OrderedPolynomial r ord n -> OrderedPolynomial r ord n

substVar n f substitutes n-th variable with polynomial f, without changing arity.

homogenize :: forall k ord n. (CoeffRing k, KnownNat n, IsMonomialOrder (n + 1) ord, IsMonomialOrder n ord) => OrderedPolynomial k ord n -> OrderedPolynomial k ord (n + 1)

Calculate the homogenized polynomial of given one, with additional variable is the last variable.

unhomogenize :: forall k ord n. (CoeffRing k, KnownNat n, IsMonomialOrder n ord, IsMonomialOrder (n + 1) ord) => OrderedPolynomial k ord (Succ n) -> OrderedPolynomial k ord n

normalize :: DecidableZero r => OrderedPolynomial r order n -> OrderedPolynomial r order n

varX :: forall r n order. (CoeffRing r, KnownNat n, IsMonomialOrder n order, (0 :< n) ~ True) => OrderedPolynomial r order n

getTerms :: OrderedPolynomial k order n -> [(k, OrderedMonomial order n)]

shiftR :: forall k r n ord. (CoeffRing r, KnownNat n, IsMonomialOrder n ord, IsMonomialOrder (k + n) ord) => SNat k -> OrderedPolynomial r ord n -> OrderedPolynomial r ord (k :+ n)

orderedBy :: OrderedPolynomial k o n -> o -> OrderedPolynomial k o n

mapCoeff :: (KnownNat n, CoeffRing b, IsMonomialOrder n ord) => (a -> b) -> OrderedPolynomial a ord n -> OrderedPolynomial b ord n

reversal :: (CoeffRing k, IsMonomialOrder 1 o) => Int -> OrderedPolynomial k o 1 -> OrderedPolynomial k o 1

padeApprox :: (Field r, DecidableUnits r, CoeffRing r, ZeroProductSemiring r, IsMonomialOrder 1 order) => Natural -> Natural -> OrderedPolynomial r order 1 -> (OrderedPolynomial r order 1, OrderedPolynomial r order 1)

eval :: (CoeffRing r, IsMonomialOrder n order, KnownNat n) => Sized n r -> OrderedPolynomial r order n -> r

Evaluate polynomial at some point.

evalUnivariate :: (CoeffRing b, IsMonomialOrder 1 order) => b -> OrderedPolynomial b order 1 -> b

substUnivariate :: (Module (Scalar r) b, Unital b, CoeffRing r, IsMonomialOrder 1 order) => b -> OrderedPolynomial r order 1 -> b

Substitute univariate polynomial using Horner's rule

minpolRecurrent :: forall k. (Eq k, ZeroProductSemiring k, DecidableUnits k, DecidableZero k, Field k) => Natural -> [k] -> Polynomial k 1

class IsOrder (n :: Nat) (ordering :: *) where

Class to lookup ordering from its (type-level) name.

Minimal complete definition

cmpMonomial

Methods

cmpMonomial :: Proxy ordering -> MonomialOrder n

newtype PadPolyL n ord poly

Constructors

PadPolyL

Fields

runPadPolyL :: OrderedPolynomial poly (Graded ord) n

padLeftPoly :: (IsMonomialOrder n ord, IsPolynomial poly) => Sing n -> ord -> poly -> PadPolyL n ord poly

module Algebra.Ring.Ideal

module Algebra.Normed

module Algebra.Internal