Algebra.Ring.Polynomial.Class

Description

This module provides abstract classes for finitary polynomial types.

Documentation

class (CoeffRing (Coefficient poly), Eq poly, DecidableZero poly, KnownNat (Arity poly), Module (Scalar (Coefficient poly)) poly, Ring poly, Commutative poly) => IsPolynomial poly where

Polynomial in terms of free associative commutative algebra generated by n-elements. To effectively compute all terms, we need monomials in addition to universality of free object.

Associated Types

type Coefficient poly :: *

Coefficient ring of polynomial type.

type Arity poly :: Nat

Arity of polynomial type.

Methods

liftMap :: (Module (Scalar (Coefficient poly)) alg, Ring alg, Commutative alg) => (Ordinal (Arity poly) -> alg) -> poly -> alg

Universal mapping for free algebra. This corresponds to the algebraic substitution operation.

subst :: (Ring alg, Commutative alg, Module (Scalar (Coefficient poly)) alg) => Sized (Arity poly) alg -> poly -> alg

A variant of liftMap, each value is given by Sized.

substWith :: (Unital r, Monoidal m) => (Coefficient poly -> r -> m) -> Sized (Arity poly) r -> poly -> m

Another variant of liftMap. This function relies on terms'; if you have more efficient implementation, it is encouraged to override this method.

sArity' :: poly -> SNat (Arity poly)

Arity of given polynomial.

sArity :: proxy poly -> SNat (Arity poly)

Arity of given polynomial, using type proxy.

arity :: proxy poly -> Integer

Non-dependent version of arity.

injectCoeff :: Coefficient poly -> poly

Inject coefficient into polynomial.

injectCoeff' :: proxy poly -> Coefficient poly -> poly

Inject coefficient into polynomial with result-type explicitly given.

monomials :: poly -> HashSet (Monomial (Arity poly))

monomials f returns the finite set of all monomials appearing in f.

terms' :: poly -> Map (Monomial (Arity poly)) (Coefficient poly)

monomials f returns the finite set of all terms appearing in f; Term is a finite map from monomials to non-zero coefficient.

coeff' :: Monomial (Arity poly) -> poly -> Coefficient poly

'coeff m f' returns the coefficient of monomial m in polynomial f.

constantTerm :: poly -> Coefficient poly

Calculates constant coefficient.

fromMonomial :: Monomial (Arity poly) -> poly

Inject monic monomial.

toPolynomial' :: (Coefficient poly, Monomial (Arity poly)) -> poly

Inject coefficient with monomial.

polynomial' :: Map (Monomial (Arity poly)) (Coefficient poly) -> poly

Construct polynomial from the given finite mapping from monomials to coefficients.

totalDegree' :: poly -> Natural

Returns total degree.

var :: Ordinal (Arity poly) -> poly

var n returns a polynomial representing n-th variable.

mapCoeff' :: (Coefficient poly -> Coefficient poly) -> poly -> poly

Adjusting coefficients of each term.

(>|*) :: Monomial (Arity poly) -> poly -> poly infixl 7

m >|* f multiplies polynomial f by monomial m.

(*|<) :: poly -> Monomial (Arity poly) -> poly infixl 7

Flipped version of (>|*)

(!*) :: Coefficient poly -> poly -> poly infixl 7

class (IsMonomialOrder (Arity poly) (MOrder poly), IsPolynomial poly) => IsOrderedPolynomial poly where

Class to lookup ordering from its (type-level) name.

Minimal complete definition

leadingTerm | leadingMonomial, leadingCoeff

Associated Types

type MOrder poly :: *

Methods

coeff :: OrderedMonomial (MOrder poly) (Arity poly) -> poly -> Coefficient poly

A variant of coeff' which takes OrderedMonomial instead of Monomial

terms :: poly -> Map (OrderedMonomial (MOrder poly) (Arity poly)) (Coefficient poly)

The default implementation is not enough efficient. So it is strongly recomended to give explicit definition to terms.

leadingTerm :: poly -> (Coefficient poly, OrderedMonomial (MOrder poly) (Arity poly))

Leading term with respect to its monomial ordering.

leadingMonomial :: poly -> OrderedMonomial (MOrder poly) (Arity poly)

Leading monomial with respect to its monomial ordering.

leadingCoeff :: poly -> Coefficient poly

Leading coefficient with respect to its monomial ordering.

orderedMonomials :: poly -> Set (OrderedMonomial (MOrder poly) (Arity poly))

The collection of all monomials in the given polynomial, with metadata of their ordering.

fromOrderedMonomial :: OrderedMonomial (MOrder poly) (Arity poly) -> poly

A variant of fromMonomial which takes OrderedMonomial as argument.

toPolynomial :: (Coefficient poly, OrderedMonomial (MOrder poly) (Arity poly)) -> poly

A variant of toPolynomial' which takes OrderedMonomial as argument.

polynomial :: Map (OrderedMonomial (MOrder poly) (Arity poly)) (Coefficient poly) -> poly

A variant of polynomial' which takes OrderedMonomial as argument.

The default implementation combines mapKeys and polynomial', hence is not enough efficient. So it is strongly recomended to give explicit definition to polynomial.

(>*) :: OrderedMonomial (MOrder poly) (Arity poly) -> poly -> poly infixl 7

A variant of '(>|*)' which takes OrderedMonomial as argument.

(*<) :: poly -> OrderedMonomial (MOrder poly) (Arity poly) -> poly infixl 7

Flipped version of (>*)

class (DecidableZero r, Ring r, Commutative r, Eq r) => CoeffRing r

Constraint synonym for rings that can be used as polynomial coefficient.

oneNorm :: (IsPolynomial poly, Normed (Coefficient poly), Monoidal (Norm (Coefficient poly))) => poly -> Norm (Coefficient poly)

1-norm of given polynomial, taking sum of norms of each coefficients.

maxNorm :: (IsPolynomial poly, Normed (Coefficient poly)) => poly -> Norm (Coefficient poly)

Maximum norm of given polynomial, taking maximum of the norms of each coefficients.

monoize :: (Field (Coefficient poly), IsOrderedPolynomial poly) => poly -> poly

Make the given polynomial monic. If the given polynomial is zero, it returns as it is.

sPolynomial :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => poly -> poly -> poly

sPolynomial calculates the S-Polynomial of given two polynomials.

pDivModPoly :: (k ~ Coefficient poly, Euclidean k, IsOrderedPolynomial poly) => poly -> poly -> (poly, poly)

pDivModPoly f g calculates the pseudo quotient and reminder of f by g.

content :: (IsPolynomial poly, Euclidean (Coefficient poly)) => poly -> Coefficient poly

The content of a polynomial f is the gcd of all its coefficients.

pp :: (Euclidean (Coefficient poly), IsPolynomial poly) => poly -> poly

pp f calculates the primitive part of given polynomial f, namely f / content(f).

injectVars :: ((Arity r :<=Arity r') ~ True, IsPolynomial r, IsPolynomial r', Coefficient r ~ Coefficient r') => r -> r'

vars :: forall poly. IsPolynomial poly => [poly]

class Show r => PrettyCoeff r where

Coefficients which admits pretty-printing

Methods

showsCoeff :: Int -> r -> ShowSCoeff

data ShowSCoeff

Pretty-printing conditional for coefficients. Each returning ShowS must not have any sign.

Constructors

Negative (MaybeShowS)

Vanished

OneCoeff

PositiveShowS

showsCoeffAsTerm :: ShowSCoeff -> ShowS

ShowS coefficients as term.

showsCoeffAsTerm Vanished "" = ""
showsCoeffAsTerm (Negative (shows "12")) "" = "-12"
showsCoeffAsTerm (Positive (shows "12")) "" = "12"

showsCoeffWithOp :: ShowSCoeff -> ShowS

ShowS coefficients prefixed with infix operator.

(shows 12 . showsCoeffWithOp Vanished) "" = "12"
(shows 12 . showsCoeffWithOp (Negative (shows 34))) "" = "12 - 34"
(shows 12 . showsCoeffWithOp (Positive (shows 34))) "" = "12 + 34"

showsPolynomialWith :: (IsPolynomial poly, PrettyCoeff (Coefficient poly)) => Sized (Arity poly) String -> Int -> poly -> ShowS

showPolynomialWith :: (IsPolynomial poly, PrettyCoeff (Coefficient poly)) => Sized (Arity poly) String -> Int -> poly -> String

Polynomial division

divModPolynomial :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => poly -> [poly] -> ([(poly, poly)], poly) infixl 7

Calculate a polynomial quotient and remainder w.r.t. second argument.

divPolynomial :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => poly -> [poly] -> [(poly, poly)] infixl 7

A Quotient of given polynomial w.r.t. the second argument.

modPolynomial :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => poly -> [poly] -> poly infixl 7

Remainder of given polynomial w.r.t. the second argument.