Algebra.Internal

Documentation

data (k :~: (a :: k)) (b :: k) :: forall k. k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k a a

withRefl :: (:~:) k a b -> ((~) k a b -> r) -> r

module Data.Proxy

toProxy :: a -> Proxy a

type Sized n a = Sized Vector n a

type Sized' n a = Sized Seq n a

coerceLength :: (n :~: m) -> Sized f n a -> Sized f m a

type SNat (n :: Nat) = Sing n

sizedLength :: ListLike (f a) a => Sized f n a -> Sing n

padVecs :: forall a n m. a -> Sized' n a -> Sized' m a -> (SNat (Max n m), Sized' (Max n m) a, Sized' (Max n m) a)

type family Flipped f a :: Nat -> Type where ...

Equations

Flipped f a = Flipped f a

pattern Flipped :: Sized f n a -> Flipped f a n

pattern OLt :: forall (t :: Nat). forall (n1 :: Nat). (n1 :< t) ~ True => Sing n1 -> Ordinal t

sNatToInt :: SNat n -> Int

Orphan instances

Hashable a => Hashable (Seq a)

Methods

hashWithSalt :: Int -> Seq a -> Int

hash :: Seq a -> Int