Algebra.Internal

Documentation

data (k :~: a) b :: forall k. k -> k -> *where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k a a

withRefl :: (:~:) k a b -> ((~) k a b -> r) -> r

Solves equality constraint without explicit coercion. It has the same effect as gcastWith, but some hacks is done to reduce runtime overhead.

module Data.Proxy

newtype WrappedField a

Constructors

WrapField

Fields

unwrapField :: a

toProxy :: a -> Proxy a

type Vector a n = SizedVector n a

type Sized n a = SizedVector n a

coerceLength :: (n :~: m) -> Sized n a -> Sized m a

type SNat n = Sing n

sizedLength :: Sized n a -> SNat n

generate :: SNat n -> (Ordinal n -> a) -> Sized n a

padVecs :: forall a n m. a -> Sized n a -> Sized m a -> (SNat (Max n m), Sized (Max n m) a, Sized (Max n m) a)

type family Flipped a :: Nat -> *where ...

Equations

Flipped a = FlippedVector a

pattern Flipped :: forall n a. Sized n a -> Flipped a n

pattern OLt :: forall t. () => forall n1. (~) Bool ((:<)Nat n1 t) True => SingNat n1 -> OrdinalNat t

sNatToInt :: SNat n -> Int