Algebra.Internal

Documentation

data (k :~: a) b :: forall k. k -> k -> *where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k a a

withRefl :: (:~:) k a b -> ((~) k a b -> r) -> r

Solves equality constraint without explicit coercion. It has the same effect as gcastWith, but some hacks is done to reduce runtime overhead.

module Data.Proxy

toProxy :: a -> Proxy a

type Sized n a = SizedVector n a

type Sized' n a = SizedSeq n a

coerceLength :: (n :~: m) -> Sized f n a -> Sized f m a

type SNat n = Sing n

sizedLength :: (PeanoOrder nat, Monomorphicable nat (Sing nat), ListLike (f a) a, Show (MonomorphicRep nat (Sing nat)), Integral (MonomorphicRep nat (Sing nat))) => Sized nat f n a -> Sing nat n

padVecs :: forall a n m. a -> Sized' n a -> Sized' m a -> (SNat (Max n m), Sized' (Max n m) a, Sized' (Max n m) a)

type family Flipped f a :: Nat -> Typewhere ...

Equations

Flipped f a = Flipped f a

pattern Flipped :: forall f n a. Sized f n a -> Flipped f a n

pattern OLt :: forall t. () => forall n1. (~) Bool ((:<)Nat n1 t) True => SingNat n1 -> OrdinalNat t

sNatToInt :: SNat n -> Int

Orphan instances

Hashable a => Hashable (Seq a)

Methods

hashWithSalt :: Int -> Seq a -> Int

hash :: Seq a -> Int