Algebra.Ring.Polynomial.Labeled

Documentation

class UniqueList xs => IsUniqueList xs

data LabPolynomial poly vars where

Constructors

LabelPolynomial :: (IsUniqueList vars, Length vars ~ Arity poly) => {..} -> LabPolynomial poly vars

Fields

unLabelPolynomial :: poly

canonicalMap :: forall xs ys poly poly'. (SingI xs, SingI ys, IsSubsetOf xs ys, Wraps xs poly, Wraps ys poly', IsPolynomial poly, IsPolynomial poly', Coefficient poly ~ Coefficient poly') => LabPolynomial poly xs -> LabPolynomial poly' ys

canonicalMap' :: (SingI xs, SingI ys, IsSubsetOf xs ys, Wraps xs poly, Wraps ys poly', IsPolynomial poly, IsPolynomial poly', Coefficient poly ~ Coefficient poly') => proxy poly' -> proxy' ys -> LabPolynomial poly xs -> LabPolynomial poly' ys

class All ((FlipSym0@@ElemSym0) @@ ys) xs ~ True => IsSubsetOf xs ys