Algebra.Ring.Polynomial.Quotient

Documentation

data Quotient r ord n ideal

The polynomial modulo the ideal indexed at the last type-parameter.

data QIdeal r ord n

reifyQuotient :: (IsMonomialOrder n ord, KnownNat n, Field r, Eq r) => Ideal (OrderedPolynomial r ord n) -> (forall ideal. Reifies ideal (QIdeal r ord n) => Proxy ideal -> a) -> a

Reifies the ideal at the type-level. The ideal can be recovered with reflect.

modIdeal :: forall ord r n ideal. (IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n), CoeffRing r, KnownNat n, Field r) => OrderedPolynomial r ord n -> Quotient r ord n ideal

Polynomial modulo ideal.

modIdeal' :: (IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n), CoeffRing r, KnownNat n, Field r) => Proxy ideal -> OrderedPolynomial r ord n -> Quotient r ord n ideal

Polynomial modulo ideal given by Proxy.

quotRepr :: Quotient r ord n ideal -> OrderedPolynomial r ord n

Representative polynomial of given quotient polynomial.

withQuotient :: (Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord) => Ideal (OrderedPolynomial r ord n) -> (forall ideal. Reifies ideal (QIdeal r ord n) => Quotient r ord n ideal) -> OrderedPolynomial r ord n

Computes polynomial modulo ideal.

vectorRep :: forall r order ideal n. (CoeffRing r, KnownNat n, IsMonomialOrder n order, Reifies ideal (QIdeal r order n)) => Quotient r order n ideal -> Vector r

genQuotVars :: forall ord n ideal r. (IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n), CoeffRing r, KnownNat n, Field r) => [Quotient r ord n ideal]

genQuotVars' :: forall ord n ideal r. (KnownNat n, Reifies ideal (QIdeal r ord n), IsMonomialOrder n ord, CoeffRing r, Field r) => Proxy ideal -> [Quotient r ord n ideal]

gBasis' :: Reifies ideal (QIdeal r ord n) => Proxy ideal -> [OrderedPolynomial r ord n]

matRep0 :: forall r ord ideal n. (Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n)) => Proxy ideal -> OrderedMonomial ord n -> Matrix r

standardMonomials :: forall ord ideal r n. (IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n), CoeffRing r, KnownNat n, Field r) => Maybe [Quotient r ord n ideal]

standardMonomials' :: (Reifies ideal (QIdeal r ord n), CoeffRing r, KnownNat n, Field r, IsMonomialOrder n ord) => Proxy ideal -> Maybe [Quotient r ord n ideal]

Find the standard monomials of the quotient ring for the zero-dimensional ideal, which are form the basis of it as k-vector space.

matRepr' :: forall r ord n ideal. (Ord r, Normed r, Field r, CoeffRing r, KnownNat n, IsMonomialOrder n ord, Reifies ideal (QIdeal r ord n)) => Quotient r ord n ideal -> Matrix r

reduce :: (Eq r, Field r, KnownNat n, IsMonomialOrder n ord) => OrderedPolynomial r ord n -> Ideal (OrderedPolynomial r ord n) -> OrderedPolynomial r ord n

Reduce polynomial modulo ideal.

multWithTable :: (Reifies ideal (QIdeal r ord n), IsMonomialOrder n ord, CoeffRing r, KnownNat n) => Quotient r ord n ideal -> Quotient r ord n ideal -> Quotient r ord n ideal

multUnamb :: (Reifies ideal (QIdeal r ord n), IsMonomialOrder n ord, CoeffRing r, KnownNat n, Field r) => Quotient r ord n ideal -> Quotient r ord n ideal -> Quotient r ord n ideal

isZeroDimensional :: (Eq r, Field r, KnownNat n, IsMonomialOrder n ord) => [OrderedPolynomial r ord n] -> Bool